requirements.txt
insightface==0.7.3
fastapi
uvicorn
faiss-cpu
opencv-python
numpy